requirements/base.txt
peewee
aiohttp
pydantic
pystemmer